LIB-31 - Added the exceptions for the deleting books feature LIB-29 - Made some improvements to the add books feature

+++ src/main/java/net/atlassian/libraryapp1/Services/BookService.java
@@ -2,6 +2,8 @@ package net.atlassian.libraryapp1.Services;
 
 import net.atlassian.libraryapp1.Exceptions.*;
 import net.atlassian.libraryapp1.Model.Book;
+import net.atlassian.libraryapp1.Model.LoggedInLibrarian;
+import net.atlassian.libraryapp1.Model.User;
 import org.dizitart.no2.Nitrite;
 import org.dizitart.no2.objects.ObjectRepository;
 import static net.atlassian.libraryapp1.Services.FileSystemService.getPathToFile;
@@ -18,14 +20,23 @@ public class BookService {
         bookRepository = database.getRepository(Book.class);
     }
 
-    public static void addBook(String title, String author, String genre, String libraryName) throws EmptyTitleFieldException, EmptyAuthorFieldException, EmptyGenreFieldException, EmptyLibraryNameFieldException {
-        checkEmptyFields(title, author, genre, libraryName);
+    public static void addBook(String title, String author, String genre) throws EmptyTitleFieldException, EmptyAuthorFieldException, EmptyGenreFieldException {
+        checkEmptyFields(title, author, genre);
         String borrowedDate="";
         String returnedDate="";
         String username="";
+        String libraryName="";
+        for(User user : UserService.userRepository.find())
+        {
+            if(user.getUsername().equals(LoggedInLibrarian.getUsername()))
+            {
+                libraryName=user.getName();
+            }
+        }
+
         bookRepository.insert(new Book(title, author, genre, libraryName, borrowedDate, returnedDate, username));
     }
-        private static void checkEmptyFields(String title, String author, String genre, String libraryName) throws EmptyTitleFieldException, EmptyAuthorFieldException, EmptyGenreFieldException, EmptyLibraryNameFieldException {
+        private static void checkEmptyFields(String title, String author, String genre) throws EmptyTitleFieldException, EmptyAuthorFieldException, EmptyGenreFieldException{
             if (title == "") {
                 throw new EmptyTitleFieldException();
             } else {
@@ -34,21 +45,25 @@ public class BookService {
                 } else {
                     if (genre == "") {
                         throw new EmptyGenreFieldException();
-                    } else {
-                        if (libraryName == "") {
-                            throw new EmptyLibraryNameFieldException();
-                        }
                     }
                 }
             }
         }
-        public static void deleteBook(String title)
+        public static void deleteBook(String title) throws WrongTitleException, EmptyTitleFieldException
         {
+            int sw=0;
             Book aux=new Book();
+            if(title=="")
+                throw new EmptyTitleFieldException();
             for(Book book : bookRepository.find())
             {
-                if(title.equals(book.getName()))
-                   aux=book;
+                if(title.equals(book.getName())) {
+                    aux = book;
+                    sw=1;
+                }
+            }
+            if(sw==0) {
+                throw new WrongTitleException();
             }
             bookRepository.remove(aux);
         }
